pyproject.toml
[build-system]
requires = ["flit_core >=3.4,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "qeapp"
dynamic = ["version"]  # read from qeapp/__init__.py
description = "Design flexible node-based workflow for AiiDA calculation."
authors = [{name = "Xing Wang", email = "xingwang1991@gmail.com"}]
readme = "README.md"
license = {file = "LICENSE"}
classifiers = [
    "Development Status :: 1 - Planning",
    "Framework :: AiiDA",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Operating System :: MacOS :: MacOS X",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering"
]
keywords = ["aiida", "workflows"]
requires-python = ">=3.9"
dependencies = [
    "ase",
    "aiida-core>=2.6",
    "fastapi",
    "uvicorn",
    "pydantic_settings",
]

[project.urls]
Documentation = "https://qeapp.readthedocs.io"
Source = "https://github.com/aiidateam/qeapp"

[project.optional-dependencies]
widget = [
    "anywidget>=0.9.11",
]
docs = [
    "sphinx_rtd_theme",
    "sphinx~=7.2",
    "sphinx-copybutton~=0.5.0",
    "sphinx-design~=0.5.0",
    "sphinx-notfound-page~=1.0",
    "sphinxext-rediraffe~=0.2.4",
    "sphinx-intl~=2.1.0",
    "sphinx-gallery",
    "myst-nb~=1.0.0",
    "nbsphinx",
]
pre-commit = [
    "pre-commit~=2.2",
    "pylint~=2.17.4",
]
tests = [
    'pgtest~=1.3',
    "pytest~=7.0",
    "pytest-cov~=2.7,<2.11",
    "playwright",
    "httpx",
]

[project.scripts]
qeapp = "qeapp.cli:cli"

[project.entry-points.'aiida.workflows']
"qeapp.workchain" = "qeapp.workflows.qeapp_workchain:QeAppWorkChain"

[tool.flit.sdist]
exclude = [
    "docs/",
    "tests/",
    "qeapp/frontend/node_modules/",
]


[tool.pylint.format]
max-line-length = 120
